Move decoder.py to models and re-arrange unitests.
