+++ deep_speech_2/utils/tests/test_error_rate.py
